Modelando e implementando a inclusão de recursos com POST.

+++ src/main/java/com/algaworks/algafood/api/controller/CozinhaController.java
@@ -1,16 +1,21 @@
 package com.algaworks.algafood.api.controller;
 
 
+import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.algaworks.algafood.domain.model.Cozinha;
@@ -43,9 +48,20 @@ public class CozinhaController {
 		
 	}
 	
+	@PostMapping
+	@ResponseStatus(HttpStatus.CREATED)
+	public Cozinha save(@RequestBody Cozinha cozinha){
+		return cozinhaRepository.save(cozinha);
+	}
+	
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteById(@PathVariable Long id){
-		cozinhaRepository.deleteById(id);
+		
+		Optional<Cozinha> cozinha = cozinhaRepository.findById(id);
+		
+		if(!cozinha.isEmpty()) {
+			cozinhaRepository.deleteById(id);
+		}
 		return ResponseEntity.noContent().build();
 	}
 }
